Update presentation to reduce flicker
There were 2 issues here: First, the scrollView's alpha when presenting in image mode started at 1, so when it was inserted as a subview, it would flash on screen instead of nicely fading in.  Second, when snapshotting a UIVibrancyEffectView, doing so after screen updates would cause the view to not be fully rendered for the screenshot.  Setting the snapshot to happen without screen updates allows the UIVibrancyEffect view to be properly captured.